% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppurple.R
\name{hapseg}
\alias{hapseg}
\title{Internal function doing simple implementation of Carter, Getz 2011 hapseg}
\usage{
hapseg(hets)
}
\arguments{
\item{hets}{data.table of genome wide het sites with $talt, $tref}
}
\description{
Phases grouped (i.e. presegmented) hets and computes cluster centers (lambdas)
and expected sufficient statistics for joint poisson
per group (i.e. segment)

Internal function
}
\author{
Marcin Imielinski
}
\keyword{internal}
