% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppurple.R
\name{ssegment}
\alias{ssegment}
\alias{log.sum.exp}
\alias{llpois}
\alias{lfactorial}
\alias{llnorml}
\alias{llnorm}
\alias{hapseg}
\alias{ppemgrid}
\alias{ppem}
\title{Internal function utilizing DNAcopy to segment a coverage profile}
\usage{
ssegment(cov, verbose = verbose)

\method{log}{sum.exp}(x)

llpois(slfx, sx, n, lambda)

lfactorial(x)

llnorm(x, sos, n, mu, sd)

hapseg(hets, verbose = FALSE)

ppemgrid(purities = NULL, ploidies = NULL, pp = NULL, segs, segs.h,
  rho = 1, rho.h = 1, k.dist = 3, K = 20, verbose = TRUE)

ppem(segs.grid, segs = NULL, segs.h = NULL,
  sd0 = segs.grid[!duplicated(j), sqrt(var(y))], sd0.k = 3,
  use.tot = TRUE, use.het = TRUE, use.uniform = FALSE, tol = 1,
  fix.sd = NULL, max_iter = 100, verbose = FALSE)
}
\arguments{
\item{verbose}{logical flag}

\item{x}{vector of log probabilities}

\item{slfx}{numeric vector of sum log factorials of data}

\item{sx}{numeric vector of sum of data}

\item{n}{numeric vector of data ns}

\item{lambda}{numeric vector of lambda parameters}

\item{sos}{numeric vector observed sum of squares}

\item{mu}{mean param of normal distribution}

\item{sd}{sigma param of normal distribution}

\item{hets}{data.table of genome wide het sites with $alt, $ref}

\item{purities}{vector of purities to sweep}

\item{ploidies}{vector of ploidies to sweep}

\item{pp}{(optional) data.table of purity x ploidy x alpha x tau for explicit combos to sweep}

\item{segs}{data.table of per segment mean coverage data, ns,  sos's across segs}

\item{segs.h}{data.table of phased cluster centers across segs}

\item{rho}{numeric scalar mean total coverage}

\item{rho.h}{numeric scalar mean het count}

\item{k.dist}{integer scalar optimization constant of how many copy states to explicitly compute probabilities for relative to the MLE value}

\item{K}{integer scalar total number of copy number states}

\item{segs.grid}{data.table collating phased het count, mean coverage, sos for every seg x purity x ploidy x copy state combos}

\item{sd0}{numeric scalar initial sd guess for total coverage clusters}

\item{sd0.k}{numeric scalar initial sd guess for distribution of integer copy number}

\item{use.tot}{logical flag of whether to use total coverage in inference}

\item{use.uniform}{logical flag of whether to use het coverage in inferendce}

\item{tol}{numeric tolerance}

\item{fix.sd}{logical flag whether to fix.sd to initial value in inference}

\item{max_iter}{integer scalar max iteration}

\item{tcov}{GRanges of binned genome-wide coverage}

\item{x}{integer vector for whiich to compute approximate log factorial}

\item{x}{numeric vector observed mean}

\item{n}{interger vector number of observations}

\item{segs}{data.table of total coverage segs inputted to ppemgrid}

\item{segs.h}{data.table of het segs inputted to ppemgrid}

\item{verbose}{logical flag}
}
\value{
GRanges of genomewise segments of piecewise constant coverage

log.sum.exp of input

log poisson joint likelihood

log factorial
}
\description{
Internal function utilizing DNAcopy to segment a coverage profile

Internal function

Internal function

Internal function

Internal function

Phases grouped (i.e. presegmented) hets and computes cluster centers (lambdas)
and expected sufficient statistics for joint poisson
per group (i.e. segment)

Internal function

Given data.tlable of segs and segs.h populated with summary stats
sets up grid across purities and ploidies and K copy states and deploys ppem on it
returns posterior probability across purities x ploidy combos given data and optimal
params

Utility function taking in
segs = grid of total copy number segments x alpha x tau x k
        --> y, mu, sd, sos, sos.k, nbins
segs.h = data.table storing grid of "high" and "low" haplotype segments x alpha x tau x k
        --> y.high, y.low, high.slfx, high.sx, low.slfx, low.sx, nbins

and returns locally optimal solution with tolerance <= tol or after max_iter iterations
initialized with sd0 and copy number prior with tau.sd = 1
the output is list with fields, each containing the following data.tables 
$pp = posterior probability of purity ploidy combo alpha tau 
$segs = grid of segments with final output
$segs.h = grid of haplotype segments with final output
$pi.k = mixing probability across total copy states for each alpha tau 
$pi.kh = mixing probability across haplotype specific copy states for each alpha tau
}
\author{
Marcin Imielinski

Marcin Imielinski

Marcin Imielinski

Marcin Imielinski

Marcin Imielinski

Marcin Imielinski

Marcin Imielinski

Marcin Imielinski
}
\keyword{internal}
