% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppurple.R
\name{lfactorial}
\alias{lfactorial}
\title{Internal function doing approximate log factorial}
\usage{
lfactorial(x)
}
\arguments{
\item{x}{integer vector for whiich to compute approximate log factorial}
}
\value{
log factorial
}
\description{
Internal function
}
\author{
Marcin Imielinski
}
