% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppurple.R
\name{llpois}
\alias{llpois}
\title{Internal function doing product of poisson log likelihoods}
\usage{
llpois(slfx, sx, n, lambda)
}
\arguments{
\item{slfx}{numeric vector of sum log factorials of data}

\item{sx}{numeric vector of sum of data}

\item{n}{numeric vector of data ns}

\item{lambda}{numeric vector of lambda parameters}
}
\value{
log poisson joint likelihood
}
\description{
Internal function
}
\author{
Marcin Imielinski
}
