% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppurple.R
\name{log.sum.exp}
\alias{log.sum.exp}
\title{Internal function doing log sum exp}
\usage{
\method{log}{sum.exp}(x)
}
\arguments{
\item{x}{vector of log probabilities}
}
\value{
log.sum.exp of input
}
\description{
Internal function
}
\author{
Marcin Imielinski
}
