% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppurple.R
\name{ppem}
\alias{ppem}
\title{ppem}
\usage{
ppem(segs.grid, segs = NULL, segs.h = NULL,
  sd0 = segs.grid[!duplicated(j), sqrt(var(y))], sd0.k = 3,
  use.tot = TRUE, use.het = TRUE, use.uniform = FALSE, tol = 1,
  fix.sd = NULL, max_iter = 100, verbose = FALSE)
}
\arguments{
\item{segs.grid}{data.table collating phased het count, mean coverage, sos for every seg x purity x ploidy x copy state combos}

\item{segs}{data.table of total coverage segs inputted to ppemgrid}

\item{segs.h}{data.table of het segs inputted to ppemgrid}

\item{sd0}{numeric scalar initial sd guess for total coverage clusters}

\item{sd0.k}{numeric scalar initial sd guess for distribution of integer copy number}

\item{use.tot}{logical flag of whether to use total coverage in inference}

\item{use.uniform}{logical flag of whether to use het coverage in inferendce}

\item{tol}{numeric tolerance}

\item{fix.sd}{logical flag whether to fix.sd to initial value in inference}

\item{max_iter}{integer scalar max iteration}

\item{verbose}{logical flag}
}
\description{
Utility function taking in
segs = grid of total copy number segments x alpha x tau x k
        --> y, mu, sd, sos, sos.k, nbins
segs.h = data.table storing grid of "high" and "low" haplotype segments x alpha x tau x k
        --> y.high, y.low, high.slfx, high.sx, low.slfx, low.sx, nbins

and returns locally optimal solution with tolerance <= tol or after max_iter iterations
initialized with sd0 and copy number prior with tau.sd = 1
the output is list with fields, each containing the following data.tables 
$pp = posterior probability of purity ploidy combo alpha tau 
$segs = grid of segments with final output
$segs.h = grid of haplotype segments with final output
$pi.k = mixing probability across total copy states for each alpha tau 
$pi.kh = mixing probability across haplotype specific copy states for each alpha tau
}
\author{
Marcin Imielinski
}
\keyword{internal}
