% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppurple.R
\name{ppemgrid}
\alias{ppemgrid}
\title{ppemgrid internal function}
\usage{
ppemgrid(purities = NULL, ploidies = NULL, pp = NULL, segs, segs.h,
  rho = 1, rho.h = 1, k.dist = 3, K = 20, verbose = TRUE)
}
\description{
Given data.tlable of segs and segs.h populated with summary stats
sets up grid across purities and ploidies and K copy states and deploys ppem on it
returns posterior probability across purities x ploidy combos given data and optimal
params
}
\author{
Marcin Imielinski
}
