% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppurple.R
\name{ppurple}
\alias{ppurple}
\alias{ppurpl}
\title{Probabilistic purity ploidy estimation}
\usage{
ppurpl(cov, hets = NULL, segs = NULL, purities = seq(0, 1, 0.1),
  ploidies = seq(1, 5, 0.2), refine = 10, K = 20, verbose = TRUE,
  min.p = 1e-04, mc.cores = 1, numchunks = mc.cores)
}
\description{
Computes posterior probability of purity ploidy for data with coverage cov (granges with value $y specifying coverage)
hets with $talt, $tref, $nref, $nalt specifying tumor and normal alt and ref allelic counts, and (optional) segments via EM algorithm. 

If segments not specified, then it is inferred from the segmentation of cov via CBS (using DNA copy)

EM begins with initial (user specified) grid of values and then refines at 10 fold (or value of refine arg, >1)
around the modes, meaning that it reruns the calculation with a finer grid completely surrounding the modes 
to output the final solution which is a data.table mapping $purity and $ploidy values
to a posterior probability $p
}
\author{
Marcin Imielinski
}
