% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppurple.R
\name{ssegment}
\alias{ssegment}
\title{Internal function utilizing DNAcopy to segment a coverage profile}
\usage{
ssegment(tcov)
}
\arguments{
\item{tcov}{GRanges of binned genome-wide coverage}
}
\value{
GRanges of genomewise segments of piecewise constant coverage
}
\description{
Internal function utilizing DNAcopy to segment a coverage profile
}
\author{
Marcin Imielinski
}
\keyword{internal}
