% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppurple.R
\name{ppemgrid}
\alias{ppemgrid}
\title{ppemgrid internal function}
\usage{
ppemgrid(purities = NULL, ploidies = NULL, pp = NULL, segs, segs.h,
  rho = 1, rho.h = 1, k.dist = 3, K = 20, verbose = TRUE)
}
\arguments{
\item{purities}{vector of purities to sweep}

\item{ploidies}{vector of ploidies to sweep}

\item{pp}{(optional) data.table of purity x ploidy x alpha x tau for explicit combos to sweep}

\item{segs}{data.table of per segment mean coverage data, ns,  sos's across segs}

\item{segs.h}{data.table of phased cluster centers across segs}

\item{rho}{numeric scalar mean total coverage}

\item{rho.h}{numeric scalar mean het count}

\item{k.dist}{integer scalar optimization constant of how many copy states to explicitly compute probabilities for relative to the MLE value}

\item{K}{integer scalar total number of copy number states}

\item{verbose}{logical flag}
}
\description{
Given data.tlable of segs and segs.h populated with summary stats
sets up grid across purities and ploidies and K copy states and deploys ppem on it
returns posterior probability across purities x ploidy combos given data and optimal
params
}
\author{
Marcin Imielinski
}
\keyword{internal}
